[#185] refactor : 코드정리, 타입 파일 컴포넌트로 이동
