+++ src/asset/icon/AlarmDot.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { alarmDot } from "../pic";
 import Img from "../../component/elem/Img";
-import { Props } from "../../model/StyleModel";
+import { Props } from "../../component/elem/StyleModel.types";
 
 const AlarmDot: React.FC<Props> = (props) => {
   return <Img {...props} src={alarmDot} />;
