+++ src/asset/icon/Cancel.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Img from "../../component/elem/Img";
-import { Props } from "../../model/StyleModel";
+import { Props } from "../../component/elem/StyleModel.types";
 import { cancel } from "../pic";
 
 const Cancel: React.FC<Props> = (props) => {
