+++ src/asset/icon/CollaboSquare.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import {  collaboIcon } from "../pic";
 import Img from "../../component/elem/Img";
-import { Props } from "../../model/StyleModel";
+import { Props } from "../../component/elem/StyleModel.types";
 
 const CollaboSquare: React.FC<Props> = (props) => {
   return <Img {...props} src={collaboIcon} />;
