+++ src/component/chat/ChatRoomItem.tsx
@@ -2,12 +2,12 @@ import React from "react";
 import styled, { keyframes } from "styled-components";
 import { more } from "../../asset/pic/index";
 import { useStomp } from "../../hook/useStomp";
-import { ChatRoom } from "../../model/ChatModel";
+import { ChatRoom } from "./ChatModel.types";
 import { useAppDispatch, useAppSelector } from "../../redux/config";
 import { isEmptyObj } from "../../util/funcs";
 import Flex, { StFlex } from "../elem/Flex";
 import Img, { ImgType } from "../elem/Img";
-import { currentRoomIdSelector, __selectChatRoom } from "./chatSlice";
+import { currentRoomIdSelector, __selectChatRoom } from "../../redux/slice/chatSlice";
 
 const ChatRoomItem: React.FC<ChatRoom> = (props) => {
   const dispatch = useAppDispatch();
