+++ src/component/elem/StyleModel.types.ts
