+++ src/component/posting/PostingFormCollabo.tsx
@@ -4,7 +4,7 @@ import { batch } from "react-redux";
 import { useParams } from "react-router-dom";
 import useInput from "../../hook/useInput";
 import useTypeModal from "../../modal/hooks/useTypeModal";
-import { CollaboForm } from "../../model/PostingModel";
+import { CollaboForm } from "./PostingModel.types";
 import { useAppDispatch, useAppSelector } from "../../redux/config";
 import { addedAudiosStateSelector } from "../../redux/slice/h5surferSlice";
 import {
