+++ src/component/posting/PostingModel.types.ts
